Rename game.php -> GameDetails.php.
